requirements.txt
apache-airflow>=2.0
requests
dbt-core
dbt-postgres